removed more unecessary code and making sure following REST principles. Multiple changes to routes, new service methods and modified where controller methods sit

+++ app/Http/Controllers/BookController.php
@@ -50,6 +50,8 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+        $this->authorize('update', $book);
+
         $this->bookService->delete($book);
 
         return back()->with('status', 'Book deleted!');
